pyproject.toml
[project]
name = "dbiz_ai_agent"
authors = [
    { name = "danls", email = "danls@dbiz.com"}
]
description = "AI Agent"
requires-python = ">=3.10"
readme = "README.md"
dynamic = ["version"]
dependencies = [
    # "frappe~=15.0.0" # Installed and managed by bench.
    "openai==2.8.1",
    "qdrant-client==1.8.2",
    "numpy==2.2.6",
    "PyPDF2==3.0.1",
    "python-docx==1.2.0",
    "python-pptx==1.0.2",
    "Pillow==11.3.0",
    "pytesseract==0.3.13",
    "pymupdf==1.26.6",
    "pdf2image==1.17.0",
    "paddleocr==3.3.2",
    "paddlex==3.3.9",
    "paddlepaddle==3.2.2",
    "requests==2.32.5",
    "openai-agents==0.4.2",
    "python-socketio==5.14.3",
    "websocket-client==1.9.0",
    "pydantic==2.11.10",
]

[build-system]
requires = ["flit_core >=3.4,<4"]
build-backend = "flit_core.buildapi"

# These dependencies are only installed when developer mode is enabled
[tool.bench.dev-dependencies]
# package_name = "~=1.1.0"
